refactor(katana): move state to backend
commit-id:af0b80f7
