+++ crates/katana/core/src/accounts.rs
@@ -17,10 +17,10 @@ use starknet_api::hash::{StarkFelt, StarkHash};
 use starknet_api::transaction::{Calldata, ContractAddressSalt};
 use starknet_api::{patricia_key, stark_felt};
 
+use crate::backend::state::{ClassRecord, MemDb, StorageRecord};
 use crate::constants::{
     DEFAULT_ACCOUNT_CONTRACT, DEFAULT_ACCOUNT_CONTRACT_CLASS_HASH, FEE_TOKEN_ADDRESS,
 };
-use crate::state::{ClassRecord, MemDb, StorageRecord};
 use crate::util::compute_legacy_class_hash;
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
