+++ crates/katana/core/src/backend/state.rs
