+++ crates/katana/core/src/lib.rs
@@ -4,7 +4,6 @@ pub mod block_context;
 pub mod constants;
 pub mod db;
 pub mod sequencer;
-pub mod state;
 pub mod util;
 
 pub mod sequencer_error;
